Enable Wayland socket access
Recent Cider releases are using Electron builds that have the last Ozone Wayland regression fixed.
As far as I can tell, this is working without issues in Sway 1.7, when adding the `--ozone-platform=wayland` flag.

Auto Ozone platform selection with the `--ozone-platform-hint=auto` flag doesn't seem to work yet,
so switching to `socket=fallback-x11` is not on the table, and in any case, there's no IME support with
Ozone Wayland.